+++ load.py
@@ -1,7 +1,7 @@
 from os import path
 
 import semantic_version
-from companion import CAPIData
+from companion import CAPIData # type: ignore
 
 import bgstally.globals
 from bgstally.bgstally import BGSTally
@@ -75,4 +75,4 @@ def dashboard_entry(cmdr:str, is_beta:bool, entry:dict) -> None:
     Handle dashboard entries
     """
     if this.state.Status.get() != CheckStates.STATE_ON: return
-    this.ui.dashboard_entry(cmdr, is_beta, entry)
+    this.dashboard_entry(cmdr, is_beta, entry)